+++ tasks/postpone.sh
@@ -15,6 +15,27 @@ postpone_task() {
     --url "https://api.todoist.com/rest/v2/tasks/$task_id" \
     --header "Authorization: Bearer ${TODOIST_API_KEY}")
   task_name=$(echo "$task_data" | jq -r '.content')
+  labels=$(echo "$task_data" | jq -r '.labels | join(", ")')
+
+  # Check if the task already has a 'Lykätty x kertaa' label
+  retry_count=1
+  if [[ "$labels" =~ "Lykätty" ]]; then
+    # Extract current retry Count
+    retry_count=$(echo "$labels" | grep -oP 'Lykätty \K[0-9]+')
+
+    # Increment the retry count
+    retry_count=$((retry_count + 1))
+  fi
+
+  # Set the correct label based on retry count
+  if [ "$retry_count" -eq 1 ]; then
+    new_label="Lykätty kerran"
+  else
+    new_label="Lykätty $retry_count kertaa"
+  fi
+
+  # Remove any existing 'Lykätty x kertaa' label from the label list
+  updated_labels=$(echo "$labels" | sed -E 's/Lykätty [0-9]+ kertaa//g' | xargs)
 
   # Check if task name contains "Google-kalenterin tapahtuma"
   if [[ "$task_name" == *"Google-kalenterin tapahtuma"* ]]; then
@@ -32,14 +53,14 @@ postpone_task() {
       --url "https://api.todoist.com/rest/v2/tasks/$task_id" \
       --header "Content-Type: application/json" \
       --header "Authorization: Bearer ${TODOIST_API_KEY}" \
-      --data "{\"due_string\": \"$current_due_string\", \"due_date\": \"$next_day\"}")
+      --data "{\"due_string\": \"$current_due_string\", \"due_date\": \"$next_day\", \"labels\": [\"$updated_labels\", \"$new_label\"]}")
   else
     # Update the task's due date
     update_response=$(curl -s --request POST \
       --url "https://api.todoist.com/rest/v2/tasks/$task_id" \
       --header "Content-Type: application/json" \
       --header "Authorization: Bearer ${TODOIST_API_KEY}" \
-      --data "{\"due_date\": \"$next_day\"}")
+      --data "{\"due_date\": \"$next_day\", \"labels\": [\"$updated_labels\", \"$new_label\"]}")
   fi
 
   # Debug