+++ tasks/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix adding tasks from Google Calendar 3 hours late on Linux
 * Fix leisure time not effective on holiday week
 * Do not postpone task if it's already scheduled
+* Add postponed count to the task label
 
 # 1.0.7: 2024-10-14